backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
greenlet
fastapi
uvicorn
sqlalchemy
aiosqlite
greenlet
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
email-validator
reportlab
slowapi
google-generativeai